Close a theft vector and two consensus divergences (third audit)
CRITICAL -- the previous fix was incomplete in a dangerous way. The guard
demands a value the signer "independently knows", but the example took that
figure from the getstaketemplate response: the same untrusted party that supplies
the coinstake. The auditor showed this is theft, not merely burning. A hostile
proposer pays 1 satoshi to the user's script (enough on its own, because
CheckBlockSignature only requires vout[1] to pay a key that can produce the block
signature, and the same-script rule applies to vault scripts, not ordinary
P2PKH), routes the whole stake to itself, and declares stake_value = 1. The guard
sees 1 >= 1 and signs. The resulting block is VALID.

The value now comes from our own view of our UTXO, and the template's claim is
cross-checked against it and refused on mismatch. A test demonstrates both
directions: trusting the proposer's figure signs the theft away, using our own
refuses it.

CONSENSUS -- an oversized nBits exponent was treated as an unbounded target, so
target_hit returned TRUE for everything. Divi calls SetCompact with a NULL
overflow flag and base_uint::operator<<= silently TRUNCATES, so a shift of >=256
yields a target of exactly ZERO and nothing can win. Backwards in the worst
direction: a relay sending absurd bits had verify_win() rubber-stamp a fabricated
win, defeating the rule that the phone verifies the relay's claim. Note the
multiply overflow is genuinely "always hit" and is unchanged -- the two overflow
paths mean opposite things and had been conflated.

CONSENSUS -- coin_age_weight clamped to zero when the timestamp preceded the
coin's start time; the node subtracts two unsigned ints, so it wraps and yields
MAXIMUM weight. Divi permits future-dated block times, so this is reachable, and
clamping made us miss stakes the network accepts. Faithfully wrapping now, bug
and all: consensus is consensus.

Also: coinstake_pays_to summed output values without saturating (panics in debug,
wraps in release on hostile input), and the script-length guard truncated a u64 to
usize, which on 32-bit Android would let a claimed length of 2^32+n pass and read
n bytes.

End-to-end re-verified after every change: still accepted, height 752.
104 tests passing.

+++ crates/lovenode-core/src/serialize.rs
@@ -94,8 +94,10 @@ impl<'a> Reader<'a> {
 
     pub fn read_var_bytes(&mut self) -> Result<Vec<u8>, String> {
         let len = self.read_compact_size()?;
-        // Guard against a hostile length claiming gigabytes.
-        if len as usize > self.remaining() {
+        // Compare in u64: `len as usize` truncates on 32-bit targets (armv7
+        // Android), so a claimed length of 2^32+n would pass this check and then
+        // read only n bytes.
+        if len > self.remaining() as u64 {
             return Err(format!("script length {len} exceeds remaining {}", self.remaining()));
         }
         Ok(self.take(len as usize)?.to_vec())
